refactor: 변경된 id 접근 수정
