+++ src/main/java/com/mobble/mobbleserver/infrastructure/persistence/article/ArticleQueryDslRepositoryImpl.java
@@ -16,7 +16,7 @@ import java.util.Set;
 import java.util.stream.Collectors;
 
 import static com.mobble.mobbleserver.domain.article.QArticle.article;
-import static com.mobble.mobbleserver.refactor.like.articleLike.entity.QArticleLike.articleLike;
+import static com.mobble.mobbleserver.domain.like.QArticleLike.articleLike;
 
 @Repository
 @RequiredArgsConstructor
@@ -40,11 +40,11 @@ public class ArticleQueryDslRepositoryImpl implements ArticleQueryDslRepository 
     public Map<Long, ArticleLikeInfoDto> findLikeInfoByArticleIdsAndMemberId(List<Long> articleIds, Long memberId) {
         List<ArticleLikeProjection> results = queryFactory
                 .select(Projections.constructor(ArticleLikeProjection.class,
-                        articleLike.article.id,
-                        articleLike.member.id
+                        articleLike.articleId,
+                        articleLike.memberId
                 ))
                 .from(articleLike)
-                .where(articleLike.article.id.in(articleIds))
+                .where(articleLike.articleId.in(articleIds))
                 .fetch();
 
         Map<Long, Integer> likeCountMap = createLikeCountMap(results);
